Test Platform: Update tructure based on NextJS best practices ( https://chat.google.com/dm/_CK8DEAAAAE/kxrV8LPA0wE/kxrV8LPA0wE?cls=10 )
